+++ app/src/main/java/com/elkins/watchlist/model/Movie.kt
@@ -20,11 +20,11 @@ data class Movie(
     val stars: String? = "N/A",
     val genres: String? = "N/A",
     val contentRating: String? = "N/A",
-    val imdbRating: String? = "N/A", // TODO Convert ratings to numeric types with custom moshi adapters
+    val imdbRating: String? = "N/A",
     val metacriticRating: String? = "N/A",
 
     var dateAdded: Date? = null,
     var haveSeen: Boolean = false,
     var following: Boolean = false,
-    var userScore: Int = 0
+    var userScore: Float = 0.0f
 ) : Parcelable
